requirements.txt
streamlit>=1.28
python-dotenv
boto3